backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
torch
torchaudio
noisereduce
transformers
accelerate
sqlalchemy
httpx
pyannote.audio

# Speed optimization - 4x faster than openai-whisper
faster-whisper

# Common deps
numpy
soundfile
openai
python-dotenv
